+++ changelog.md
@@ -1,3 +1,4 @@
 ### General
 
+- Added carpet rule `fluidNoPushPlayer`
 - `betterOpPlayerNoCheat` now effects `/clone` as well